url and cleaning code
