+++ package.json
@@ -10,8 +10,8 @@
     "@types/node": "^12.12.62",
     "@types/react": "^16.9.49",
     "@types/react-dom": "^16.9.8",
-    "@types/styled-components": "^5.1.3",
     "@types/react-router-dom": "^5.1.5",
+    "@types/styled-components": "^5.1.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
