+++ src/styles/services.css
@@ -1,3 +1,5 @@
+/*THIS FILE CAN BE DELETED WHEN THE HOME PAGE HAS BEEN DONE*/
+
 div {
   display: flex;
   flex-direction: column;
